pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "multi-agent"
version = "0.6.0"
description = "Multi-agent orchestration framework for IDE-based AI coding assistants"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.4,<1",
    "langgraph-checkpoint-sqlite>=2,<3",
    "pydantic>=2.5,<3",
    "pyyaml>=6,<7",
    "jinja2>=3.1,<4",
    "click>=8,<9",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0",
    "pytest-cov>=5.0",
    "ruff>=0.3",
    "mypy>=1.8",
]
mcp = [
    "mcp>=1,<2",
]

[project.scripts]
ma = "multi_agent.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/multi_agent"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
